fix: setUserCounter not being correctly debounced
This occured because server actions are instantiated in the server as needed
so the cannot be debounced as usual. To fix this, I encountered a solution
to debounce react states, and use this to call setUserCounter instead.
Also, lodash is not used anymore, so the dependency was removed.
Also, a high severity vulnerability was fixed.